Add atomic run once label
After projectatomic/atomic#952 it's now possible to add a label to an image
atomic.run=once so that when installed as a system container it will be run
immediately with runc (no systemd involved) and then uninstalled.

This use case matches containerized Ansible playbooks very well, so we add
that label in playbook2image so that the usage experience of packaged
playbooks run as system containers matches standalone container runs.